+++ server/migrations/20240131093638_create_comment.up.sql
@@ -9,7 +9,7 @@ create table comments
     target_id uuid not null,
     content varchar(200) not null,
     target_type target_types not null,
-    likes smallint not null default 0,
+    likes bigint not null default 0,
     created_at timestamp (6) with time zone default now() not null,
     updated_at timestamp (6) with time zone default now() not null
 );
